Clean FontFaceHandle interface header
